Fix Update Apartment
This script currently only updates the label of the apt. from the config however does not pass the id through, thus is an apartment gets updated in the db it would just be "Integrity Way" instead of "Integrity Way 9032". Passing the label through is questionable as you can get it from the type but not the focus of this fix.
NOTE: this only works with 9 or less apartments in the config as it is currently ripping the id from the apartment29032 id in the db, ideally the id of the apartment would be stored individually in the database to future proof this.