Update main.lua
Quick fix to the doorbell side of things, current implementation does not close when there is nobody home as the apartment variable has a initialized state of {} from the server, thus it will never be nil. 
For this to work properly the function should be checking if the list is empty rather than nil. 

Works in game: https://imgur.com/a/ijUkCbX
- Have you personally loaded this code into an updated qbcore project and checked all it's functionality? [yes] (Be honest)
- Does your code fit the style guidelines? [yes]
- Does your PR fit the contribution guidelines? [yes]